12 - Reorganized hooks for profile page, new tab for profile data changes
