+++ messages/ukr.json
@@ -188,6 +188,36 @@
   "profile": {
     "title": "Мій профіль",
     "subtitle": "Привіт, {name}! Тут ви можете керувати даними свого профілю.",
+    "tabs": {
+      "professional": "Професійні дані",
+      "personal": "Особисті дані"
+    },
+    "personal": {
+      "title": "Особисті дані",
+      "subtitle": "Керуйте своїми основними особистими даними та контактною інформацією",
+      "form": {
+        "first_name": "Ім'я",
+        "last_name": "Прізвище",
+        "email": "Електронна пошта",
+        "phone": "Телефон",
+        "email_note": "Електронну адресу не можна змінити з міркувань безпеки",
+        "save": "Зберегти дані",
+        "saving": "Збереження..."
+      },
+      "placeholders": {
+        "first_name": "Марія",
+        "last_name": "Петренко",
+        "email": "maria@example.com",
+        "phone": "+372 1234 5678"
+      },
+      "messages": {
+        "save_success": "Особисті дані успішно оновлено!",
+        "save_error": "Помилка при збереженні особистих даних",
+        "loading": "Завантаження даних...",
+        "name_required": "Ім'я та прізвище є обов'язковими",
+        "email_required": "Електронна пошта є обов'язковою"
+      }
+    },
     "form": {
       "title": "Мої дані",
       "subtitle": "Заповніть дані свого профілю, щоб консультанти та наставники могли отримати кращий огляд ваших навичок та пройденого шляху. Ці дані видимі лише лікарям Medunacy.",
@@ -209,7 +239,10 @@
     "messages": {
       "save_success": "Дані успішно збережено!",
       "save_error": "Помилка при збереженні даних",
-      "loading": "Завантаження даних..."
+      "loading": "Завантаження даних...",
+      "university_required": "Університет/Школа є обов'язковим полем",
+      "specialization_required": "Спеціальність/Напрямок є обов'язковим полем",
+      "languages_required": "Необхідно вибрати принаймні одну мову"
     }
   }
 }
