+++ src/components/profile/ProfilePage.tsx
@@ -5,7 +5,7 @@ import { useEffect } from 'react';
 import { useRouter, usePathname } from 'next/navigation';
 import Header from '../layout/Header';
 import ProfileHero from './ProfileHero';
-import ProfileForm from './ProfileForm';
+import ProfileTabs from './ProfileTabs/ProfileTabs';
 import { AuthModalProvider } from '@/contexts/AuthModalContext';
 import { useAuth } from '@/contexts/AuthContext';
 
@@ -47,7 +47,7 @@ export default function ProfilePage() {
         {/* Main Profile Content */}
         <main>
           <ProfileHero />
-          <ProfileForm />
+          <ProfileTabs />
         </main>
       </div>
     </AuthModalProvider>
