+++ src/email-templates/reset-password-et.html
@@ -1,0 +1,110 @@
+<!DOCTYPE html>
+<html lang="et">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Lähtesta oma Medunacy parool</title>
+</head>
+<body style="margin: 0; padding: 0; font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif; line-height: 1.6; color: #333; background-color: #FBF6E9;">
+    <table width="100%" cellpadding="0" cellspacing="0" style="background-color: #FBF6E9;">
+        <tr>
+            <td align="center" style="padding: 20px;">
+                <table width="600" cellpadding="0" cellspacing="0" style="background: white; border-radius: 16px; overflow: hidden; box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);">
+                    <!-- Header -->
+                    <tr>
+                        <td style="background: linear-gradient(135deg, #118B50 0%, #5DB996 100%); padding: 40px 30px; text-align: center;">
+                            <div style="font-size: 32px; margin-bottom: 12px;">🔒</div>
+                            <h1 style="color: white; font-size: 32px; font-weight: 700; margin: 0 0 8px 0;">Medunacy</h1>
+                            <p style="color: #E3F0AF; font-size: 16px; margin: 0;">Turvalisus on meie prioriteet</p>
+                        </td>
+                    </tr>
+                    
+                    <!-- Content -->
+                    <tr>
+                        <td style="padding: 40px 30px;">
+                            <h2 style="font-size: 24px; font-weight: 700; color: #118B50; margin: 0 0 16px 0;">Parooli lähtestamine</h2>
+                            
+                            <p style="font-size: 16px; margin: 0 0 20px 0; color: #555;">
+                                Saime sinu parooli lähtestamise taotluse. Kui sina seda taotlust ei esitanud, võid selle e-kirja lihtsalt ignoreerida.
+                            </p>
+
+                            <!-- Warning Box -->
+                            <table width="100%" cellpadding="0" cellspacing="0" style="margin: 20px 0;">
+                                <tr>
+                                    <td style="background: #FFF4E5; border: 1px solid #FFD093; border-radius: 8px; padding: 16px;">
+                                        <p style="margin: 0; font-size: 14px; color: #92400E;">
+                                            <span style="color: #F59E0B; font-size: 20px; vertical-align: middle;">⚠️</span>
+                                            <strong>Tähelepanu:</strong> See link kehtib ainult 1 tunni jooksul turvalisuse tagamiseks.
+                                        </p>
+                                    </td>
+                                </tr>
+                            </table>
+                            
+                            <!-- CTA Button -->
+                            <table width="100%" cellpadding="0" cellspacing="0" style="margin: 32px 0;">
+                                <tr>
+                                    <td align="center">
+                                        <a href="{{ .ResetURL }}" style="display: inline-block; background: #DC2626; color: #ffffff; text-decoration: none; padding: 16px 32px; border-radius: 12px; font-weight: 600; font-size: 16px;">
+                                            Lähtesta parool
+                                        </a>
+                                    </td>
+                                </tr>
+                            </table>
+
+                            <!-- Info Box -->
+                            <table width="100%" cellpadding="0" cellspacing="0" style="margin: 20px 0;">
+                                <tr>
+                                    <td style="background: #F0FDF4; border: 1px solid #BBF7D0; border-radius: 8px; padding: 16px;">
+                                        <p style="margin: 0; font-size: 14px; color: #047857;">
+                                            <span style="color: #059669; font-size: 20px; vertical-align: middle;">ℹ️</span>
+                                            Pärast parooli lähtestamist logitakse sind automaatselt välja kõigist seadmetest.
+                                        </p>
+                                    </td>
+                                </tr>
+                            </table>
+                            
+                            <p style="font-size: 16px; margin: 20px 0 10px 0; color: #555;">
+                                <strong>Turvalisuse soovitused:</strong>
+                            </p>
+                            <ul style="color: #555; margin-left: 20px; padding-left: 0;">
+                                <li>Kasuta tugevat parooli, mis sisaldab vähemalt 8 tähemärki</li>
+                                <li>Lisa numbreid, suurtähti ja erimärke</li>
+                                <li>Ära kasuta sama parooli mitmes kohas</li>
+                            </ul>
+                            
+                            <!-- Alternative Link -->
+                            <table width="100%" cellpadding="0" cellspacing="0" style="margin: 24px 0;">
+                                <tr>
+                                    <td style="background: #F8F9FA; padding: 20px; border-radius: 8px; border-left: 4px solid #5DB996;">
+                                        <p style="margin: 0 0 8px 0; font-size: 14px; color: #666;">
+                                            <strong>Link ei tööta?</strong>
+                                        </p>
+                                        <p style="margin: 0 0 8px 0; font-size: 14px; color: #666;">
+                                            Kopeeri ja kleebi see aadress oma veebilehitseja aadressiribale:
+                                        </p>
+                                        <p style="background: #E9ECEF; padding: 8px 12px; border-radius: 4px; font-family: 'Courier New', monospace; font-size: 12px; word-break: break-all; margin: 8px 0 0 0; color: #495057;">
+                                            {{ .ResetURL }}
+                                        </p>
+                                    </td>
+                                </tr>
+                            </table>
+                        </td>
+                    </tr>
+                    
+                    <!-- Footer -->
+                    <tr>
+                        <td style="background: #F8F9FA; padding: 30px; text-align: center; border-top: 1px solid #E9ECEF;">
+                            <p style="font-size: 14px; color: #6C757D; margin: 0 0 8px 0;">
+                                See e-kiri saadeti, kuna taotleti parooli lähtestamist.
+                            </p>
+                            <p style="font-size: 14px; color: #6C757D; margin: 0;">
+                                <strong style="color: #118B50;">Medunacy</strong> - Turvalisus ja privaatsus on meie prioriteedid
+                            </p>
+                        </td>
+                    </tr>
+                </table>
+            </td>
+        </tr>
+    </table>
+</body>
+</html>
