+++ src/email-templates/reset-password-ukr.html
@@ -1,0 +1,110 @@
+<!DOCTYPE html>
+<html lang="uk">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Скинути пароль Medunacy</title>
+</head>
+<body style="margin: 0; padding: 0; font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif; line-height: 1.6; color: #333; background-color: #FBF6E9;">
+    <table width="100%" cellpadding="0" cellspacing="0" style="background-color: #FBF6E9;">
+        <tr>
+            <td align="center" style="padding: 20px;">
+                <table width="600" cellpadding="0" cellspacing="0" style="background: white; border-radius: 16px; overflow: hidden; box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);">
+                    <!-- Header -->
+                    <tr>
+                        <td style="background: linear-gradient(135deg, #118B50 0%, #5DB996 100%); padding: 40px 30px; text-align: center;">
+                            <div style="font-size: 32px; margin-bottom: 12px;">🔒</div>
+                            <h1 style="color: white; font-size: 32px; font-weight: 700; margin: 0 0 8px 0;">Medunacy</h1>
+                            <p style="color: #E3F0AF; font-size: 16px; margin: 0;">Безпека - наш пріоритет</p>
+                        </td>
+                    </tr>
+                    
+                    <!-- Content -->
+                    <tr>
+                        <td style="padding: 40px 30px;">
+                            <h2 style="font-size: 24px; font-weight: 700; color: #118B50; margin: 0 0 16px 0;">Скидання пароля</h2>
+                            
+                            <p style="font-size: 16px; margin: 0 0 20px 0; color: #555;">
+                                Ми отримали ваш запит на скидання пароля. Якщо ви не надсилали цей запит, просто проігноруйте цей лист.
+                            </p>
+
+                            <!-- Warning Box -->
+                            <table width="100%" cellpadding="0" cellspacing="0" style="margin: 20px 0;">
+                                <tr>
+                                    <td style="background: #FFF4E5; border: 1px solid #FFD093; border-radius: 8px; padding: 16px;">
+                                        <p style="margin: 0; font-size: 14px; color: #92400E;">
+                                            <span style="color: #F59E0B; font-size: 20px; vertical-align: middle;">⚠️</span>
+                                            <strong>Увага:</strong> Це посилання дійсне лише протягом 1 години для забезпечення безпеки.
+                                        </p>
+                                    </td>
+                                </tr>
+                            </table>
+                            
+                            <!-- CTA Button -->
+                            <table width="100%" cellpadding="0" cellspacing="0" style="margin: 32px 0;">
+                                <tr>
+                                    <td align="center">
+                                        <a href="{{ .ResetURL }}" style="display: inline-block; background: #DC2626; color: #ffffff; text-decoration: none; padding: 16px 32px; border-radius: 12px; font-weight: 600; font-size: 16px;">
+                                            Скинути пароль
+                                        </a>
+                                    </td>
+                                </tr>
+                            </table>
+
+                            <!-- Info Box -->
+                            <table width="100%" cellpadding="0" cellspacing="0" style="margin: 20px 0;">
+                                <tr>
+                                    <td style="background: #F0FDF4; border: 1px solid #BBF7D0; border-radius: 8px; padding: 16px;">
+                                        <p style="margin: 0; font-size: 14px; color: #047857;">
+                                            <span style="color: #059669; font-size: 20px; vertical-align: middle;">ℹ️</span>
+                                            Після скидання пароля ви будете автоматично виведені з усіх пристроїв.
+                                        </p>
+                                    </td>
+                                </tr>
+                            </table>
+                            
+                            <p style="font-size: 16px; margin: 20px 0 10px 0; color: #555;">
+                                <strong>Рекомендації з безпеки:</strong>
+                            </p>
+                            <ul style="color: #555; margin-left: 20px; padding-left: 0;">
+                                <li>Використовуйте надійний пароль з мінімум 8 символів</li>
+                                <li>Додайте цифри, великі літери та спеціальні символи</li>
+                                <li>Не використовуйте один пароль у кількох місцях</li>
+                            </ul>
+                            
+                            <!-- Alternative Link -->
+                            <table width="100%" cellpadding="0" cellspacing="0" style="margin: 24px 0;">
+                                <tr>
+                                    <td style="background: #F8F9FA; padding: 20px; border-radius: 8px; border-left: 4px solid #5DB996;">
+                                        <p style="margin: 0 0 8px 0; font-size: 14px; color: #666;">
+                                            <strong>Посилання не працює?</strong>
+                                        </p>
+                                        <p style="margin: 0 0 8px 0; font-size: 14px; color: #666;">
+                                            Скопіюйте та вставте цю адресу в адресний рядок вашого браузера:
+                                        </p>
+                                        <p style="background: #E9ECEF; padding: 8px 12px; border-radius: 4px; font-family: 'Courier New', monospace; font-size: 12px; word-break: break-all; margin: 8px 0 0 0; color: #495057;">
+                                            {{ .ResetURL }}
+                                        </p>
+                                    </td>
+                                </tr>
+                            </table>
+                        </td>
+                    </tr>
+                    
+                    <!-- Footer -->
+                    <tr>
+                        <td style="background: #F8F9FA; padding: 30px; text-align: center; border-top: 1px solid #E9ECEF;">
+                            <p style="font-size: 14px; color: #6C757D; margin: 0 0 8px 0;">
+                                Цей лист був надісланий, оскільки було запитано скидання пароля.
+                            </p>
+                            <p style="font-size: 14px; color: #6C757D; margin: 0;">
+                                <strong style="color: #118B50;">Medunacy</strong> - Безпека та конфіденційність - наші пріоритети
+                            </p>
+                        </td>
+                    </tr>
+                </table>
+            </td>
+        </tr>
+    </table>
+</body>
+</html>
